Add code for ""PDF Processing Pipeline" module. The prompt used was:
I would like you to now start generating code in Python. Follow the best Python coding practices in terms of code, module and file layout. As a first step, generate code for the part "PDF Processing Pipeline". The pdf files will be available in a GCP bucket. The code should have the ability to scan the bucket when triggered and find files that have not yet been processed. Those files should be converted to markdown and saved to Firestore. Generate an example main function that illustrates the functionality of this module. The overall application is named "recall-assistant".